HttpRequest: Fix "Implicit conversion from float to int loses precision". Remove obsolete SelectOptionsRenderer::acceptMultipleSelections and use SelectOptionsField::acceptMultipleSelections instead. Fix "session_destroy(): Session object destruction failed".

+++ core/HttpRequest.php
@@ -154,7 +154,7 @@ class HttpRequest
 		foreach ($acceptedLanguages as $acceptedLanguageDetails) {
 			$languageAndQuality = explode(separator: ';q=', string: $acceptedLanguageDetails);
 			$languageCode = trim(string: explode(separator: '-', string: $languageAndQuality[0])[0]);
-			$quality = (array_key_exists(key: 1, array: $languageAndQuality) ? round(num: (float)$languageAndQuality[1], precision: 2) : 1.0) * 100;
+			$quality = (int)((array_key_exists(key: 1, array: $languageAndQuality) ? round(num: (float)$languageAndQuality[1], precision: 2) : 1.0) * 100);
 			if (!array_key_exists(key: $quality, array: $listByQuality)) {
 				$listByQuality[$quality] = $languageCode;
 			}
